--- examples/deploy/addon/helloworld-cloudevents/kustomization.yaml
@@ -1,28 +1,0 @@
-namespace: open-cluster-management
-
-resources:
-- resources/cluster_role.yaml
-- resources/cluster_role_binding.yaml
-- resources/service_account.yaml
-- resources/managed_clusterset_binding.yaml
-- resources/placement.yaml
-- resources/addon_deployment_config.yaml
-- resources/helloworld_cloudevents_clustermanagementaddon.yaml
-- resources/helloworld_cloudevents_controller.yaml
-- resources/work-driver-config.yaml
-
-images:
-- name: quay.io/open-cluster-management/addon-examples
-  newName: quay.io/open-cluster-management/addon-examples
-  newTag: latest
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-
-vars:
-  - name: EXAMPLE_IMAGE_NAME
-    objref:
-      apiVersion: apps/v1
-      kind: Deployment
-      name: helloworldcloudevents-controller
-    fieldref:
-      fieldpath: spec.template.spec.containers.[name=helloworldcloudevents-controller].image
