--- examples/deploy/addon/helloworld-cloudevents/resources/addon_deployment_config.yaml
@@ -1,6 +1,0 @@
-apiVersion: addon.open-cluster-management.io/v1alpha1
-kind: AddOnDeploymentConfig
-metadata:
-  name: global
-spec:
-  agentInstallNamespace: open-cluster-management-agent-addon
