--- examples/deploy/addon/helloworld-cloudevents/resources/cluster_role.yaml
@@ -1,50 +1,0 @@
-  kind: ClusterRole
-  apiVersion: rbac.authorization.k8s.io/v1
-  metadata:
-    name: helloworldcloudevents-addon
-  rules:
-    - apiGroups: [""]
-      resources: ["configmaps", "events"]
-      verbs: ["get", "list", "watch", "create", "update", "delete", "deletecollection", "patch"]
-    - apiGroups: ["coordination.k8s.io"]
-      resources: ["leases"]
-      verbs: ["get", "list", "watch", "create", "update", "patch"]
-    - apiGroups: ["rbac.authorization.k8s.io"]
-      resources: ["roles", "rolebindings"]
-      verbs: ["get", "list", "watch", "create", "update", "delete"]
-    - apiGroups: ["authorization.k8s.io"]
-      resources: ["subjectaccessreviews"]
-      verbs: ["get", "create"]
-    - apiGroups: ["certificates.k8s.io"]
-      resources: ["certificatesigningrequests", "certificatesigningrequests/approval"]
-      verbs: ["get", "list", "watch", "create", "update"]
-    - apiGroups: ["certificates.k8s.io"]
-      resources: ["signers"]
-      verbs: ["approve"]
-    - apiGroups: ["cluster.open-cluster-management.io"]
-      resources: ["managedclusters"]
-      verbs: ["get", "list", "watch"]
-    - apiGroups: ["work.open-cluster-management.io"]
-      resources: ["manifestworks"]
-      verbs: ["create", "update", "get", "list", "watch", "delete", "deletecollection", "patch"]
-    - apiGroups: ["addon.open-cluster-management.io"]
-      resources: ["managedclusteraddons/finalizers"]
-      verbs: ["update"]
-    - apiGroups: [ "addon.open-cluster-management.io" ]
-      resources: [ "clustermanagementaddons/finalizers" ]
-      verbs: [ "update" ]
-    - apiGroups: [ "addon.open-cluster-management.io" ]
-      resources: [ "clustermanagementaddons/status" ]
-      verbs: ["update", "patch"]
-    - apiGroups: ["addon.open-cluster-management.io"]
-      resources: ["clustermanagementaddons"]
-      verbs: ["get", "list", "watch", "patch"]
-    - apiGroups: ["addon.open-cluster-management.io"]
-      resources: ["managedclusteraddons"]
-      verbs: ["get", "list", "watch", "create", "update", "delete"]
-    - apiGroups: ["addon.open-cluster-management.io"]
-      resources: ["managedclusteraddons/status"]
-      verbs: ["update", "patch"]
-    - apiGroups: ["addon.open-cluster-management.io"]
-      resources: ["addondeploymentconfigs"]
-      verbs: ["get", "list", "watch"]
