--- examples/deploy/addon/helloworld-cloudevents/resources/cluster_role_binding.yaml
@@ -1,12 +1,0 @@
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: helloworldcloudevents-addon
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: helloworldcloudevents-addon
-subjects:
-  - kind: ServiceAccount
-    name: helloworldcloudevents-addon-sa
-    namespace: open-cluster-management
