--- examples/deploy/addon/helloworld-cloudevents/resources/helloworld_cloudevents_clustermanagementaddon.yaml
@@ -1,21 +1,0 @@
-apiVersion: addon.open-cluster-management.io/v1alpha1
-kind: ClusterManagementAddOn
-metadata:
-  name: helloworldcloudevents
-spec:
-  addOnMeta:
-    displayName: helloworldcloudevents
-    description: "helloworldcloudevents is an example addon using cloudevents created by go template"
-  supportedConfigs:
-  - group: addon.open-cluster-management.io
-    resource: addondeploymentconfigs
-  installStrategy:
-    type: Placements
-    placements:
-      - name: global
-        namespace: open-cluster-management
-        configs:
-          - group: addon.open-cluster-management.io
-            resource: addondeploymentconfigs
-            name: global
-            namespace: open-cluster-management
