--- examples/deploy/addon/helloworld-cloudevents/resources/managed_clusterset_binding.yaml
@@ -1,6 +1,0 @@
-apiVersion: cluster.open-cluster-management.io/v1beta2
-kind: ManagedClusterSetBinding
-metadata:
-  name: global
-spec:
-  clusterSet: global
