--- examples/deploy/addon/helloworld-cloudevents/resources/placement.yaml
@@ -1,10 +1,0 @@
-apiVersion: cluster.open-cluster-management.io/v1beta1
-kind: Placement
-metadata:
-  name: global
-spec:
-  clusterSets:
-    - global
-  predicates:
-    - requiredClusterSelector:
-        labelSelector: {}
