--- examples/deploy/addon/helloworld-cloudevents/resources/service_account.yaml
@@ -1,4 +1,0 @@
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: helloworldcloudevents-addon-sa
