--- examples/deploy/addon/helloworld-cloudevents/resources/work-driver-config.yaml
@@ -1,11 +1,0 @@
-apiVersion: v1
-kind: Secret
-metadata:
-  name: work-driver-config
-stringData:
-  config.yaml: |
-    brokerHost: mosquitto.mqtt:1883
-    topics:
-      sourceEvents: sources/addon-manager/clusters/+/sourceevents
-      agentEvents: sources/addon-manager/clusters/+/agentevents
-      sourceBroadcast: sources/addon-manager/sourcebroadcast
