--- examples/helloworld_cloudevents/helloworld.go
@@ -1,95 +1,0 @@
-package helloworld_cloudevents
-
-import (
-	"embed"
-	"fmt"
-	"os"
-
-	"k8s.io/client-go/rest"
-	"open-cluster-management.io/addon-framework/examples/rbac"
-	"open-cluster-management.io/addon-framework/pkg/addonfactory"
-	"open-cluster-management.io/addon-framework/pkg/agent"
-	"open-cluster-management.io/addon-framework/pkg/utils"
-	addonapiv1alpha1 "open-cluster-management.io/api/addon/v1alpha1"
-	clusterv1 "open-cluster-management.io/api/cluster/v1"
-	workapiv1 "open-cluster-management.io/api/work/v1"
-)
-
-const (
-	DefaultImage          = "quay.io/open-cluster-management/addon-examples:latest"
-	AddonName             = "helloworldcloudevents"
-	InstallationNamespace = "open-cluster-management-agent-addon"
-)
-
-//go:embed manifests
-//go:embed manifests/templates
-var FS embed.FS
-
-func NewRegistrationOption(kubeConfig *rest.Config, addonName, agentName string) *agent.RegistrationOption {
-	return &agent.RegistrationOption{
-		CSRConfigurations: agent.KubeClientSignerConfigurations(addonName, agentName),
-		CSRApproveCheck:   utils.DefaultCSRApprover(agentName),
-		PermissionConfig:  rbac.AddonRBAC(kubeConfig),
-	}
-}
-
-func GetDefaultValues(cluster *clusterv1.ManagedCluster,
-	addon *addonapiv1alpha1.ManagedClusterAddOn) (addonfactory.Values, error) {
-
-	image := os.Getenv("EXAMPLE_IMAGE_NAME")
-	if len(image) == 0 {
-		image = DefaultImage
-	}
-
-	manifestConfig := struct {
-		KubeConfigSecret string
-		ClusterName      string
-		Image            string
-	}{
-		KubeConfigSecret: fmt.Sprintf("%s-hub-kubeconfig", addon.Name),
-		ClusterName:      cluster.Name,
-		Image:            image,
-	}
-
-	return addonfactory.StructToValues(manifestConfig), nil
-}
-
-func AgentHealthProber() *agent.HealthProber {
-	return &agent.HealthProber{
-		Type: agent.HealthProberTypeWork,
-		WorkProber: &agent.WorkHealthProber{
-			ProbeFields: []agent.ProbeField{
-				{
-					ResourceIdentifier: workapiv1.ResourceIdentifier{
-						Group:     "apps",
-						Resource:  "deployments",
-						Name:      "helloworldcloudevents-agent",
-						Namespace: InstallationNamespace,
-					},
-					ProbeRules: []workapiv1.FeedbackRule{
-						{
-							Type: workapiv1.WellKnownStatusType,
-						},
-					},
-				},
-			},
-			HealthCheck: func(identifier workapiv1.ResourceIdentifier, result workapiv1.StatusFeedbackResult) error {
-				if len(result.Values) == 0 {
-					return fmt.Errorf("no values are probed for deployment %s/%s", identifier.Namespace, identifier.Name)
-				}
-				for _, value := range result.Values {
-					if value.Name != "ReadyReplicas" {
-						continue
-					}
-
-					if *value.Value.Integer >= 1 {
-						return nil
-					}
-
-					return fmt.Errorf("readyReplica is %d for deployement %s/%s", *value.Value.Integer, identifier.Namespace, identifier.Name)
-				}
-				return fmt.Errorf("readyReplica is not probed")
-			},
-		},
-	}
-}
