--- examples/helloworld_cloudevents/manifests/templates/clusterrolebinding.yaml
@@ -1,12 +1,0 @@
-kind: ClusterRoleBinding
-apiVersion: rbac.authorization.k8s.io/v1
-metadata:
-  name: helloworldcloudevents-agent
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: cluster-admin
-subjects:
-  - kind: ServiceAccount
-    name: helloworldcloudevents-agent-sa
-    namespace: {{ .AddonInstallNamespace }}
