--- examples/helloworld_cloudevents/manifests/templates/deployment.yaml
@@ -1,68 +1,0 @@
-kind: Deployment
-apiVersion: apps/v1
-metadata:
-  name: helloworldcloudevents-agent
-  namespace: {{ .AddonInstallNamespace }}
-  labels:
-    app: helloworldcloudevents-agent
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: helloworldcloudevents-agent
-  template:
-    metadata:
-      labels:
-        app: helloworldcloudevents-agent
-    spec:
-      serviceAccountName: helloworldcloudevents-agent-sa
-{{- if .NodeSelector }}
-      nodeSelector:
-      {{- range $key, $value := .NodeSelector }}
-        "{{ $key }}": "{{ $value }}"
-      {{- end }}
-{{- end }}
-{{- if .Tolerations }}
-      tolerations:
-      {{- range $toleration := .Tolerations }}
-      - key: "{{ $toleration.Key }}"
-        value: "{{ $toleration.Value }}"
-        effect: "{{ $toleration.Effect }}"
-        operator: "{{ $toleration.Operator }}"
-        {{- if $toleration.TolerationSeconds }}
-        tolerationSeconds: {{ $toleration.TolerationSeconds }}
-        {{- end }}
-      {{- end }}
-{{- end }}
-      volumes:
-      - name: hub-config
-        secret:
-          secretName: {{ .KubeConfigSecret }}
-      containers:
-      - name: helloworldcloudevents-agent
-        image: {{ .Image }}
-        imagePullPolicy: IfNotPresent
-{{- if or .HTTPProxy .HTTPSProxy}}
-        env:
-        {{- if .HTTPProxy }}
-          - name: HTTP_PROXY
-            value: {{ .HTTPProxy }}
-        {{- end }}
-        {{- if .HTTPSProxy }}
-          - name: HTTPS_PROXY
-            value: {{ .HTTPSProxy }}
-        {{- end }}
-        {{- if .NoProxy }}
-          - name: NO_PROXY
-            value: {{ .NoProxy }}
-        {{- end }}
-{{- end }}
-        args:
-          - "/helloworld_cloudevents"
-          - "agent"
-          - "--hub-kubeconfig=/var/run/hub/kubeconfig"
-          - "--cluster-name={{ .ClusterName }}"
-          - "--addon-namespace={{ .AddonInstallNamespace }}"
-        volumeMounts:
-          - name: hub-config
-            mountPath: /var/run/hub
