--- examples/helloworld_cloudevents/manifests/templates/serviceaccount.yaml
@@ -1,5 +1,0 @@
-kind: ServiceAccount
-apiVersion: v1
-metadata:
-  name: helloworldcloudevents-agent-sa
-  namespace: {{ .AddonInstallNamespace }}
