+++ pkg/addonmanager/cloudevents/manager.go
@@ -19,15 +19,7 @@ import (
 	workinformers "open-cluster-management.io/api/client/work/informers/externalversions"
 
 	"open-cluster-management.io/addon-framework/pkg/addonmanager"
-	"open-cluster-management.io/addon-framework/pkg/addonmanager/controllers/addonconfig"
-	"open-cluster-management.io/addon-framework/pkg/addonmanager/controllers/agentdeploy"
-	"open-cluster-management.io/addon-framework/pkg/addonmanager/controllers/certificate"
-	"open-cluster-management.io/addon-framework/pkg/addonmanager/controllers/cmaconfig"
-	"open-cluster-management.io/addon-framework/pkg/addonmanager/controllers/cmamanagedby"
-	"open-cluster-management.io/addon-framework/pkg/addonmanager/controllers/registration"
-	"open-cluster-management.io/addon-framework/pkg/basecontroller/factory"
 	"open-cluster-management.io/addon-framework/pkg/index"
-	"open-cluster-management.io/addon-framework/pkg/utils"
 	cloudeventswork "open-cluster-management.io/sdk-go/pkg/cloudevents/work"
 	"open-cluster-management.io/sdk-go/pkg/cloudevents/work/source/codec"
 )
@@ -42,8 +34,6 @@ type cloudeventsAddonManager struct {
 func (a *cloudeventsAddonManager) Start(ctx context.Context) error {
 	config := a.GetConfig()
 	addonAgents := a.GetAddonAgents()
-	addonConfigs := a.GetAddonConfigs()
-	syncContexts := a.GetSyncContexts()
 
 	var addonNames []string
 	for key := range addonAgents {
@@ -162,112 +152,9 @@ func (a *cloudeventsAddonManager) Start(ctx context.Context) error {
 		return err
 	}
 
-	v1CSRSupported, v1beta1Supported, err := utils.IsCSRSupported(kubeClient)
+	err = a.StartWithInformers(ctx, workClient, workInformers, kubeInformers, addonInformers, clusterInformers, dynamicInformers)
 	if err != nil {
 		return err
-	}
-
-	for _, agentImpl := range addonAgents {
-		for _, configGVR := range agentImpl.GetAgentAddonOptions().SupportedConfigGVRs {
-			addonConfigs[configGVR] = true
-		}
-	}
-
-	deployController := agentdeploy.NewAddonDeployController(
-		workClient,
-		addonClient,
-		clusterInformers.Cluster().V1().ManagedClusters(),
-		addonInformers.Addon().V1alpha1().ManagedClusterAddOns(),
-		workInformers,
-		addonAgents,
-	)
-
-	registrationController := registration.NewAddonRegistrationController(
-		addonClient,
-		clusterInformers.Cluster().V1().ManagedClusters(),
-		addonInformers.Addon().V1alpha1().ManagedClusterAddOns(),
-		addonAgents,
-	)
-
-	// This controller is used during migrating addons to be managed by addon-manager.
-	// This should be removed when the migration is done.
-	// The migration plan refer to https://github.com/open-cluster-management-io/ocm/issues/355.
-	managementAddonController := cmamanagedby.NewCMAManagedByController(
-		addonClient,
-		addonInformers.Addon().V1alpha1().ClusterManagementAddOns(),
-		addonAgents,
-		utils.FilterByAddonName(addonAgents),
-	)
-
-	var addonConfigController, managementAddonConfigController factory.Controller
-	if len(addonConfigs) != 0 {
-		addonConfigController = addonconfig.NewAddonConfigController(
-			addonClient,
-			addonInformers.Addon().V1alpha1().ManagedClusterAddOns(),
-			addonInformers.Addon().V1alpha1().ClusterManagementAddOns(),
-			dynamicInformers,
-			addonConfigs,
-			utils.FilterByAddonName(addonAgents),
-		)
-		managementAddonConfigController = cmaconfig.NewCMAConfigController(
-			addonClient,
-			addonInformers.Addon().V1alpha1().ClusterManagementAddOns(),
-			dynamicInformers,
-			addonConfigs,
-			utils.FilterByAddonName(addonAgents),
-		)
-	}
-
-	var csrApproveController factory.Controller
-	var csrSignController factory.Controller
-	// Spawn the following controllers only if v1 CSR api is supported in the
-	// hub cluster. Under v1beta1 CSR api, all the CSR objects will be signed
-	// by the kube-controller-manager so custom CSR controller should be
-	// disabled to avoid conflict.
-	if v1CSRSupported {
-		csrApproveController = certificate.NewCSRApprovingController(
-			kubeClient,
-			clusterInformers.Cluster().V1().ManagedClusters(),
-			kubeInformers.Certificates().V1().CertificateSigningRequests(),
-			nil,
-			addonInformers.Addon().V1alpha1().ManagedClusterAddOns(),
-			addonAgents,
-		)
-		csrSignController = certificate.NewCSRSignController(
-			kubeClient,
-			clusterInformers.Cluster().V1().ManagedClusters(),
-			kubeInformers.Certificates().V1().CertificateSigningRequests(),
-			addonInformers.Addon().V1alpha1().ManagedClusterAddOns(),
-			addonAgents,
-		)
-	} else if v1beta1Supported {
-		csrApproveController = certificate.NewCSRApprovingController(
-			kubeClient,
-			clusterInformers.Cluster().V1().ManagedClusters(),
-			nil,
-			kubeInformers.Certificates().V1beta1().CertificateSigningRequests(),
-			addonInformers.Addon().V1alpha1().ManagedClusterAddOns(),
-			addonAgents,
-		)
-	}
-
-	a.SetSyncContexts(append(syncContexts, deployController.SyncContext()))
-
-	go deployController.Run(ctx, 1)
-	go registrationController.Run(ctx, 1)
-	go managementAddonController.Run(ctx, 1)
-
-	if addonConfigController != nil {
-		go addonConfigController.Run(ctx, 1)
-	}
-	if managementAddonConfigController != nil {
-		go managementAddonConfigController.Run(ctx, 1)
-	}
-	if csrApproveController != nil {
-		go csrApproveController.Run(ctx, 1)
-	}
-	if csrSignController != nil {
-		go csrSignController.Run(ctx, 1)
 	}
 
 	kubeInformers.Start(ctx.Done())
